orabug #30226232: incorrect behaviour of chained CRUD ops in webserver mode - CRUD Collection.Find: fix wrong management of raw zvals which was the main reason of incorrect behaviour (in Collection_modify there was kept zval* object_zv;) - refactorings + apply util::zvalue

+++ util/value.cc
@@ -507,7 +507,10 @@ void zvalue::clear()
 void zvalue::reserve(std::size_t size)
 {
 	if (is_array()) {
-		zend_hash_extend(Z_ARRVAL(zv), size, (Z_ARRVAL(zv)->u.flags & HASH_FLAG_PACKED));
+		zend_hash_extend(
+			Z_ARRVAL(zv),
+			static_cast<uint32_t>(size),
+			(Z_ARRVAL(zv)->u.flags & HASH_FLAG_PACKED));
 	} else {
 		zval_ptr_dtor(&zv);
 		array_init_size(&zv, static_cast<uint32_t>(size));
